zombienet_orchestrator

Module pjs_helper

Enums

ReturnValue

Js script return value, mapping types that can be deserialized as serde_json::Value or not, for the latest an string is returned witht the error message.

Functions

pjs_build_template

pjs_exec

Type Aliases

PjsResult

pjs-rs success Result type